% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fast_Test_aSPA.R
\name{Fast_Test_aSPA}
\alias{Fast_Test_aSPA}
\title{Fast Test average Superior Predictive Ability}
\usage{
Fast_Test_aSPA(
  LossDiff,
  weights = NULL,
  L,
  B = 999,
  num_cores = 1,
  seed = stats::runif(1, 0, .Machine$integer.max)
)
}
\arguments{
\item{LossDiff}{the T x H matrix forecast path loss differential}

\item{weights}{the 1 x H vector of weights for the losses at different horizons. For instance \code{weights <- matlab::ones(1,20)/20}}

\item{L}{integer, the parameter for the moving block bootstrap}

\item{B}{integer, the number of bootstrap iterations. Default 999}

\item{num_cores}{integer, the number of cores to be used. Default 1}

\item{seed}{integer, random seed for replication}
}
\value{
A list containing two objects:
\item{"p_value"}{the p-value for aSPA}
\item{"t_aSPA"}{the statistics for aSPA}
}
\description{
Implements the test for average Superior Predictive Ability (aSPA) of Quaedvlieg (2021)
}
\examples{
## Test for aSPA and uSPA
Trow <- 200 
H <- 12
Mmethods <- 5
weights <- rep(1/H,H)
Losses <- matrix(rnorm(Trow*H, mean = 0), nrow = Trow, ncol = H)

Fast_Test_aSPA(LossDiff=Losses, weights=weights, L=3, B=10)

}
\references{
Quaedvlieg, Rogier. "Multi-horizon forecast comparison." Journal of Business & Economic Statistics 39.1 (2021): 40-53.
}
\seealso{
\code{\link{Test_uSPA}}
}
\author{
Luca Barbaglia \url{https://lucabarbaglia.github.io/}
}
