% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fast_Test_uSPA.R
\name{Fast_Test_uSPA}
\alias{Fast_Test_uSPA}
\title{Fast Test uniform Superior Predictive Ability}
\usage{
Fast_Test_uSPA(
  LossDiff,
  L,
  B = 999,
  num_cores = 1,
  seed = runif(1, 0, .Machine$integer.max)
)
}
\arguments{
\item{LossDiff}{the T x H matrix forecast path loss differential}

\item{L}{the parameter for the moving block bootstrap}

\item{B}{integer, the number of bootstrap iterations. Default 999}
}
\value{
A list containing two objects:
\item{"p_value"}{the p-value for uSPA}
\item{"t_uSPA"}{the statistics for uSPA}
}
\description{
Implements the test for uniform Superior Predictive Ability (uSPA) of Quaedvlieg (2021)
}
\examples{
## Test for uSPA
Trow <- 200 
H <- 12
Mmethods <- 5
Losses <- matrix(rnorm(Trow*H, mean = 0), nrow = Trow, ncol = H)

Fast_Test_uSPA(LossDiff=Losses, L=3, B=10)

}
\references{
Quaedvlieg, Rogier. "Multi-horizon forecast comparison." Journal of Business & Economic Statistics 39.1 (2021): 40-53.
}
\seealso{
\code{\link{Test_aSPA}}
}
\author{
Luca Barbaglia \url{https://lucabarbaglia.github.io/}
}
