% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LossDiff_aSPA.R
\docType{data}
\name{LossDiff_aSPA}
\alias{LossDiff_aSPA}
\title{Example for aSPA}
\format{
A matrix TxH. Each column is a loss-function for horizon h=1,..H.
}
\usage{
data(LossDiff_aSPA)
}
\description{
A set of loss-differences for horizons H=1,...,20, that has aSPA, but not uSPA.
}
\references{
Quaedvlieg, Rogier. "Multi-horizon forecast comparison." Journal of Business & Economic Statistics 39.1 (2021): 40-53.
}
\seealso{
\code{\link{Test_aSPA}} to test for aSPA.
}
\keyword{datasets}
