% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LossDiff_uSPA.R
\docType{data}
\name{LossDiff_uSPA}
\alias{LossDiff_uSPA}
\title{Example for uSPA}
\format{
A matrix TxH. Each column is a loss-function for horizon h=1,..H.
}
\usage{
data(LossDiff_uSPA)
}
\description{
A set of loss-differences for horizons H=1,...,20, that has both aSPA and uSPA.
}
\references{
Quaedvlieg, Rogier. "Multi-horizon forecast comparison." Journal of Business & Economic Statistics 39.1 (2021): 40-53.
}
\seealso{
\code{\link{Test_uSPA}} to test for uSPA.
}
\keyword{datasets}
