% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Test_aSPA.R
\name{Test_aSPA}
\alias{Test_aSPA}
\title{Test average Superior Predictive Ability}
\usage{
Test_aSPA(LossDiff, weights, L)
}
\arguments{
\item{LossDiff}{the TxH forecast path loss differential}

\item{weights}{the 1xH vector of weights for the losses at different horizons. For instance \code{weights <- matlab::ones(1,20)/20}}

\item{L}{the parameter for the moving block bootstrap}
}
\value{
A list containg two objects:
\item{"p_value"}{the p-value for aSPA}
\item{"t_aSPA"}{the statistics for aSPA}
}
\description{
Implements the test for average Superior Predictive Ability (aSPA) of Quaedvlieg (2021)
}
\examples{
## Test for aSPA and uSPA
data(LossDiff_aSPA)
weights <- matlab::ones(1,20)/20
Test_aSPA(LossDiff=LossDiff_aSPA, weights=weights, L=3)
Test_uSPA(LossDiff=LossDiff_aSPA, L=3)

}
\references{
Quaedvlieg, Rogier. "Multi-horizon forecast comparison." Journal of Business & Economic Statistics 39.1 (2021): 40-53.
}
\seealso{
\code{\link{Test_uSPA}}
}
