% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_function.R
\name{test_function}
\alias{test_function}
\title{test_function}
\usage{
test_function(seed)
}
\arguments{
\item{seed}{A real number between 0 and 1 that determines the splitting probability in the prior (which is used as the importance sampler of tree models). Quadrianto and Ghahramani (2015) recommend a value less than 0.5 .}
}
\value{
A number
}
\description{
test_function
}
\examples{



}
